Make `PioneerAssert` entry point for all custom assertions
